Fix compatibility issues of numba and torch.meshgrid().
