+++ __init__.py
@@ -1,8 +1,20 @@
 
 from .planar_as_base import IDENTITY_ROT
+from . import planar_as_base
 
-from . import camera_model_sampler
-from . import full_view_rotation
-from . import six_images
+# Camera model based sampler.
+from .camera_model_sampler import CameraModelRotation
 
+# Sampler that works with Unreal Engine's cube renderings.
+from .full_view_rotation import FullViewRotation
+
+# Six image sampler. The PyTorch version.
+from .six_images import (SixPlanarTorch, dummy_debug_callback)
+
+# Six image sampler. The numba version.
+from .numba_support_check import is_numba_supported
+if is_numba_supported():
+    from .six_images_py_numba import SixPlanarNumba
+
+# All samplers in a dict.
 from .register import (SAMPLERS, make_object)
